private-cluster-tailscale/infra/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=7.0.0,<8.0.0
pulumi-kubernetes>=4.0.0,<5.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
